requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.3
Babel==2.14.0
billiard==3.6.4.0
captcha==0.4
celery==4.3.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.7
cssselect==1.2.0
Django==3.2.23
django-haystack==3.2.1
django-redis==5.4.0
django-widget-tweaks==1.4.12
dnspython==1.16.0
docxcompose==1.4.0
docxtpl==0.16.8
easyocr==1.7.1
eventlet==0.25.1
greenlet==3.0.3
idna==3.7
imageio==2.31.2
importlib-metadata==4.13.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==3.1.3
jsons==1.6.3
kombu==4.6.11
lxml==5.2.2
MarkupSafe==2.1.3
monotonic==1.6
networkx==2.6.3
ninja==1.11.1.1
numpy==1.21.6
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pyclipper==1.3.0.post5
pycparser==2.21
PyMySQL==1.1.0
pyquery==2.0.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
shapely==2.0.4
six==1.16.0
sqlparse==0.4.4
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.7.1
typish==1.9.3
urllib3==2.0.7
vine==1.3.0
Whoosh==2.7.4
zipp==3.15.0
